requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
SQLAlchemy==2.0.46
asyncpg==0.29.0
python-dotenv==1.2.1
httpx==0.27.0